requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chainlit==0.7.700
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.100.1
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
frozenlist==1.4.1
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.75
Lazify==0.4.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
playsound==1.3.0
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.8.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.1
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.9.0
uptrace==1.22.0
urllib3==2.1.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
